Connected front with back

+++ autoAttendanceMonitoring/models.py
@@ -68,7 +68,7 @@ class Student(models.Model):
     FName = models.CharField(max_length=50)
     LName = models.CharField(max_length=50)
     year_of_education = models.ForeignKey(YearOfEducation, models.DO_NOTHING)
-    statistics = models.DecimalField(decimal_places=3, default=None, max_digits=4)
+    statistics = models.IntegerField(default=None)
 
 
 class Subject(models.Model):
@@ -83,7 +83,7 @@ class Lesson(models.Model):
     end_time = models.DateTimeField()
     kind = models.CharField(max_length=10)
     subject = models.ForeignKey(Subject, models.DO_NOTHING)
-    statistics = models.DecimalField(decimal_places=3, default=None, max_digits=4)
+    statistics = models.IntegerField(default=None)
 
 
 class IsTaughtBy(models.Model):
